apps/api/pyproject.toml
[tool.poetry]
name = "scr-api"
version = "0.1.0"
description = "SCR Platform — FastAPI Backend"
authors = ["SCR Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = "^2.10.3"
pydantic-settings = "^2.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.1"
httpx = "^0.28.1"
boto3 = "^1.35.86"
structlog = "^24.4.0"
python-multipart = "^0.0.19"
elasticsearch = {extras = ["async"], version = "^8.17.0"}
tenacity = "^9.0.0"
orjson = "^3.10.12"
psycopg2-binary = "^2.9.11"
svix = "^1.86.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
ruff = "^0.8.4"
mypy = "^1.13.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
